# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'utilisateur",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails de l'utilisateur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Utilisateur non trouvé"
          }
        },
        "summary": "Récupérer un utilisateur par son ID",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UsersController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'utilisateur",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Utilisateur mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Utilisateur non trouvé"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Mettre à jour un utilisateur",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'utilisateur",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Utilisateur supprimé"
          },
          "404": {
            "description": "Utilisateur non trouvé"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Supprimer un utilisateur",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/lessons": {
      "get": {
        "operationId": "LessonsController_getAllLessons",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des leçons",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Obtenir toutes les leçons publiées",
        "tags": [
          "Lessons"
        ]
      },
      "post": {
        "operationId": "LessonsController_createLesson",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Leçon créée"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Créer une nouvelle leçon (instructeur)",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/api/lessons/{id}": {
      "get": {
        "operationId": "LessonsController_getLessonById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la leçon",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails de la leçon",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonResponseDto"
                }
              }
            }
          }
        },
        "summary": "Obtenir une leçon par ID",
        "tags": [
          "Lessons"
        ]
      },
      "put": {
        "operationId": "LessonsController_updateLesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la leçon",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLessonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Leçon mise à jour"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Mettre à jour une leçon",
        "tags": [
          "Lessons"
        ]
      },
      "delete": {
        "operationId": "LessonsController_deleteLesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la leçon",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leçon supprimée"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Supprimer une leçon",
        "tags": [
          "Lessons"
        ]
      }
    },
...
    "/api/admin/dashboard": {
      "get": {
        "operationId": "AdminController_getDashboardStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir les statistiques du dashboard",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/users": {
      "get": {
        "operationId": "AdminController_getAllUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir tous les utilisateurs",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "operationId": "AdminController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir un utilisateur par ID",
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Supprimer un utilisateur",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/users/{id}/admin": {
      "put": {
        "operationId": "AdminController_makeUserAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Promouvoir/Rétrograder un utilisateur admin",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/users/{id}/publisher": {
      "put": {
        "operationId": "AdminController_updatePublisherStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePublisherStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Activer/Désactiver le statut publisher",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/applications": {
      "get": {
        "operationId": "AdminController_getAllApplications",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir toutes les candidatures",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/applications/{id}/approve": {
      "put": {
        "operationId": "AdminController_approveApplication",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveApplicationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Approuver/Modifier le statut d'une candidature",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/lessons": {
      "get": {
        "operationId": "AdminController_getAllLessons",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir toutes les leçons",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/lessons/{id}": {
      "delete": {
        "operationId": "AdminController_deleteLesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Supprimer une leçon",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/blog-posts": {
      "get": {
        "operationId": "AdminController_getAllBlogPosts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir tous les articles de blog",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/blog-posts/{id}": {
      "delete": {
        "operationId": "AdminController_deleteBlogPost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Supprimer un article de blog",
        "tags": [
          "Admin"
        ]
      }
    }
  },
  "info": {
    "title": "EduLearn API",
    "description": "API pour la plateforme EduLearn",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "avatar": {
            "type": "string",
            "example": "https://api.dicebear.com/7.x/avataaars/svg?seed=john"
          },
          "isPublisher": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-10-25T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "avatar",
          "isPublisher",
          "createdAt"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "john_doe_updated",
            "description": "Nouveau nom d'utilisateur"
          },
          "email": {
            "type": "string",
            "example": "newemail@example.com",
            "description": "Nouvelle adresse email"
          },
          "password": {
            "type": "string",
            "example": "NewPassword123!",
            "description": "Nouveau mot de passe"
          },
          "avatar": {
            "type": "string",
            "example": "https://api.dicebear.com/7.x/avataaars/svg?seed=newavatar",
            "description": "URL du nouvel avatar"
          },
          "isPublisher": {
            "type": "boolean",
            "example": true,
            "description": "Statut publisher"
          }
        }
      },
      "InstructorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "username": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "avatar"
        ]
      },
      "LessonResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "thumbnail": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "isFree": {
            "type": "boolean"
          },
          "level": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ]
          },
          "duration": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "enrollmentCount": {
            "type": "number"
          },
          "instructor": {
            "$ref": "#/components/schemas/InstructorDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "isEnrolled": {
            "type": "boolean"
          },
          "progress": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "thumbnail",
          "price",
          "isFree",
          "level",
          "status",
          "duration",
          "tags",
          "enrollmentCount",
          "instructor",
          "createdAt",
          "isEnrolled",
          "progress"
        ]
      },
      "CreateLessonDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Introduction à JavaScript"
          },
          "description": {
            "type": "string",
            "example": "Apprenez les bases de JavaScript"
          },
          "thumbnail": {
            "type": "string",
            "example": "https://example.com/thumbnail.jpg"
          },
          "price": {
            "type": "number",
            "example": 29.99
          },
          "isFree": {
            "type": "boolean",
            "example": false
          },
          "level": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "example": "beginner"
          },
          "duration": {
            "type": "number",
            "example": 120
          },
          "tags": {
            "example": [
              "javascript",
              "programming"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "description",
          "price",
          "isFree",
          "level",
          "duration"
        ]
      },
      "UpdateLessonDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Introduction à JavaScript"
          },
          "description": {
            "type": "string",
            "example": "Apprenez les bases de JavaScript"
          },
          "thumbnail": {
            "type": "string",
            "example": "https://example.com/thumbnail.jpg"
          },
          "price": {
            "type": "number",
            "example": 29.99
          },
          "isFree": {
            "type": "boolean",
            "example": false
          },
          "level": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "example": "beginner"
          },
          "duration": {
            "type": "number",
            "example": 120
          },
          "tags": {
            "example": [
              "javascript",
              "programming"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "isAdmin": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "isAdmin"
        ]
      },
      "UpdatePublisherStatusDto": {
        "type": "object",
        "properties": {
          "isPublisher": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "isPublisher"
        ]
      },
      "ApproveApplicationDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "reviewing",
              "shortlisted",
              "interview",
              "accepted",
              "rejected",
              "withdrawn"
            ]
          },
          "feedback": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      }
    }
  }
}